GTK 4 : 36 avertissements a 17, dont deux vrais bugs et un correctif jamais porte
Le port GTK 4 compilait avec 36 avertissements. Il en reste 17, et ce qui reste
est volontaire — voir la fin de ce message.

## Trois defauts de comportement

tag_attributes.c — « if (!paramspec->flags & G_PARAM_WRITABLE) ». Le « ! » lie
plus fort que le « & » : l'expression vaut (!flags) & 2, donc TOUJOURS 0. Le
garde-fou « propriete non modifiable » ne s'est jamais declenche. Le port GTK 3
porte ce correctif depuis longtemps, commentaire a l'appui ; le port GTK 4 ne
l'avait jamais recu.

glade_support.c — le « return TRUE; » etait hors du « if », malgre son
indentation. La fonction renvoyait donc vrai des la premiere iteration de la
boucle, sans egard au resultat de la comparaison : seul signals[0] pouvait etre
connecte, et l'echec etait annonce comme un succes. Accolades posees.
⚠ Le port GTK 3 a le meme defaut, non corrige ici.

widget_tree.c — dans widget_tree_save(), « text » n'est affectee qu'a
l'interieur de la boucle sur les colonnes. Si cette boucle ne tourne pas, « text »
etait lue par fprintf puis passee a g_free() sans avoir jamais ete initialisee.
On utilise « line », qui vaut g_strdup("") avant la boucle et recoit « text » a
chaque tour : meme resultat quand la boucle tourne, plus de lecture indeterminee
quand elle ne tourne pas. Cela supprime aussi une fuite de « line ».
⚠ Le port GTK 3 a la meme construction, non corrigee ici.

## Le reste

- variables.c : quatre « else » ambigus dans _tree_insert et _tree_find. Le
  liage etait deja celui que l'indentation suggerait : accolades explicites,
  aucun changement de sens.
- variables.c : le parametre « window_id » masquait le global du meme nom.
  Renomme en « drop_window_id » — et les TROIS usages du corps avec lui. Les
  renommer etait obligatoire : sans cela, la comparaison qui decide quels
  widgets sont detruits (l.1467) serait passee du parametre au global.
- widget_tree.c : deux locales masquaient les globaux icon_name et stock_id du
  meme fichier. Prefixees « cell_ ».
- automaton.c : widget_moved n'est appelee que depuis le bloc GtkSocket, compile
  hors du binaire en GTK 4. Elle passe sous la meme garde #ifndef GTK4_NO_SOCKET
  plutot que d'etre supprimee — elle reste vraie pour X11. Cela retire aussi le
  masquage du global « window ».
- Six comparaisons signe/non-signe (automaton.c, signals.c, stringman.c) : le
  size_t est ramene dans le domaine signe des autres operandes. Cast du size_t
  vers gint et non l'inverse : si l'operande signe etait negatif, le convertir en
  size_t donnerait une valeur enorme et inverserait le test.
- gtkdialog_lexer.l : « %option noinput ». YY_INPUT est redefini plus haut et lit
  par getnextchar(), donc input() ne sert jamais ; flex l'engendrait quand meme.
  Pas de « nounput » : unput() EST utilisee, l.773.

stringman.c est encode en ISO-8859-1, pas en UTF-8. Il a ete edite en octets
pour ne pas abimer ses accents ; « file -i » le confirme inchange.

## Ce qui reste, et pourquoi

15 des 17 avertissements restants sont des « variable affectee mais jamais
utilisee ». Ce ne sont pas des variables mortes : ce sont des valeurs que le code
calcule correctement, et qu'une macro de gtk4-compat.h jette. Exemple :
gtk_container_set_border_width vaut ((void)0), donc border-width est ignore ;
gtk_file_chooser_button_new(title, action) construit gtk_button_new_with_label
(title), donc <filechooser> est un bouton ordinaire.

Les faire taire effacerait la seule alerte qui subsiste sur ces fonctions
mortes. Elles restent. Un audit separe les recense.

Les deux derniers : « statement with no effect » sur
gtk_window_set_icon_from_file, qui vaut (TRUE) — une macro qui annonce un succes
sans rien faire, meme raison de la garder visible ; et process_string, definie et
jamais appelee, a l'identique dans les deux ports.

## Mesures

Reconstruction complete, GCC 15.3.0 / GTK 4.22.4 : 67 fichiers, 0 erreur,
36 -> 17 avertissements.
tests/xml/run_tests.sh all : 55 PASS | 0 FAIL sur chacun des deux binaires.
Temoin hors suite, lexer regenere : un XML avec attributs et imbrication rend
8 lignes d'IR, rc=0.

+++ gtk4sermo/gtk4sermo_1.1.1/src/signals.c
@@ -1539,9 +1539,9 @@ gboolean widget_signal_executor_eval_condition(gchar *condition)
 			}
 
 			/* Fix any escaped double-quotes if present */
-			for (index = 0; index < strlen(argument); index++)
+			for (index = 0; index < (gint)strlen(argument); index++)
 				if (argument[index] == '\\' && argument[index + 1] == '"' )
-					for (count = index; count < strlen(argument); count++)
+					for (count = index; count < (gint)strlen(argument); count++)
 						argument[count] = argument[count + 1];
 
 			/* Remove enveloping spaces if present */
